+++ xtask/src/main.rs
@@ -697,11 +697,10 @@ fn lint_packages(workspace: &Path, args: LintPackagesArgs) -> Result<()> {
                 }
 
                 Package::EspWifi => {
-                    let mut features =
-                        format!("--features={chip},async,ps-min-modem,defmt,dump-packets,sys-logs");
+                    let mut features = format!("--features={chip},defmt,sys-logs");
 
                     if device.contains("wifi") {
-                        features.push_str(",wifi-default,esp-now,embassy-net,sniffer")
+                        features.push_str(",wifi-default,esp-now,sniffer")
                     }
                     if device.contains("bt") {
                         features.push_str(",ble")